PublicFunction을 RealmHelper로 변경. 구현을 오버로드로 수정 realm 전역 변수로 사용하던것을 원래대로 복구
