+++ Project/PayRoad/RealmHelper.swift
@@ -1,0 +1,72 @@
+//
+//  RealmHelper.swift
+//  PayRoad
+//
+//  Created by Yoo Seok Kim on 2017. 8. 9..
+//  Copyright © 2017년 REFUEL. All rights reserved.
+//
+
+import UIKit
+
+import RealmSwift
+
+class RealmHelper {
+    class func tableViewNotificationToken<T>(for tableView: UITableView, list: List<T>) -> NotificationToken {
+        // Observe Results Notifications
+        let closure: (RealmCollectionChange<AnyRealmCollection<T>>) -> Void = { (changes: RealmCollectionChange) in
+            switch changes {
+            case .initial:
+                // Results are now populated and can be accessed without blocking the UI
+                tableView.reloadData()
+                break
+            case .update(_, let deletions, let insertions, let modifications):
+                // Query results have changed, so apply them to the UITableView
+                tableView.beginUpdates()
+                tableView.insertRows(at: insertions.map({ IndexPath(row: $0, section: 0)}),
+                                     with: .automatic)
+                tableView.deleteRows(at: deletions.map({ IndexPath(row: $0, section: 0)}),
+                                     with: .automatic)
+                tableView.reloadRows(at: modifications.map({ IndexPath(row: $0, section: 0)}),
+                                     with: .automatic)
+                tableView.endUpdates()
+                break
+            case .error(let error):
+                // An error occurred while opening the Realm file on the background worker thread
+                fatalError("\(error)")
+                break
+            }
+        }
+        
+        return list._addNotificationBlock(closure)
+    }
+    
+    class func tableViewNotificationToken<T>(for tableView: UITableView, results: Results<T>) -> NotificationToken {
+        // Observe Results Notifications
+        let closure: (RealmCollectionChange<AnyRealmCollection<T>>) -> Void = { (changes: RealmCollectionChange) in
+            switch changes {
+            case .initial:
+                // Results are now populated and can be accessed without blocking the UI
+                tableView.reloadData()
+                break
+            case .update(_, let deletions, let insertions, let modifications):
+                // Query results have changed, so apply them to the UITableView
+                tableView.beginUpdates()
+                tableView.insertRows(at: insertions.map({ IndexPath(row: $0, section: 0)}),
+                                     with: .automatic)
+                tableView.deleteRows(at: deletions.map({ IndexPath(row: $0, section: 0)}),
+                                     with: .automatic)
+                tableView.reloadRows(at: modifications.map({ IndexPath(row: $0, section: 0)}),
+                                     with: .automatic)
+                tableView.endUpdates()
+                break
+            case .error(let error):
+                // An error occurred while opening the Realm file on the background worker thread
+                fatalError("\(error)")
+                break
+            }
+        }
+        
+        return results._addNotificationBlock(closure)
+    }
+    
+}
