+++ Project/PayRoad/TravelEditorViewController.swift
@@ -16,6 +16,8 @@ class TravelEditorViewController: UIViewController {
     @IBOutlet weak var startDateTextField: UITextField!
     @IBOutlet weak var endDateTextField: UITextField!
     
+    let realm = try! Realm()
+    
     override func viewDidLoad() {
         super.viewDidLoad()
